+++ App/ui/src/constants.js
@@ -1,2 +1,2 @@
-export const APP_VERSION = 'v2.7.3'
+export const APP_VERSION = 'v2.7.4'
 export const CHANGELOG_URL = 'https://github.com/robinsondan87/GeekyThingsProductCatalogue/blob/main/CHANGELOG.md'